abstraction(IDetail) for details(DIP and OCP)

+++ CarService/Storage/DetailDTO.cs
@@ -10,8 +10,8 @@ namespace CarService.Storage
     {
         public string Name {  get; set; }
         public string Model {  get; set; }
-        public int Count {  get; set; }
-        public int Price {  get; set; }
+        public double Count {  get; set; }
+        public double Price {  get; set; }
         public string Value {  get; set; }
 
     }
